refactor(archtest): 迁移 mqtt 规则族 → module-path-agnostic Check* + 缩 baseline [#1636]
MOVE mqtt_funnel_test.go（MQTT-CLIENT-ID-NAMESPACE-01 / MQTT-TOPIC-NAMESPACE-01）+
mqtt_reason_redaction_test.go（REASON-NAME-REDACTION-01）的检测逻辑到非 test 的 mqtt_funnel.go /
mqtt_reason_redaction.go（Check* + scanner helpers）；平台符号路径经 PlatformModulePath 派生
（mqttPkgPath / mqttErrcodePkgPath const + A1ScannerFires 自检的 reflect fixture f.PkgPath）。

共享符号 mqttPkgPath / mqttEnclosingFuncName / mqttEnclosingFuncAllowed 从 _test.go 迁入 mqtt_funnel.go
（同包，5 个兄弟 mqtt test 文件 — callsite/dlx/reason/connect-deadline/config-validate — 不改即编译通过，
无重复声明）。_test.go 瘦身为 dogfood Report() + 保留 A1/A2 ScannerFires / A2RedFixture / BlindSpot /
NonVacuousness / ScannerNonVacuous 自检。两文件 Tags: FlatNonDefaultTags() → Check* 读 cfg.BuildTags，
dogfood 传 BuildTags: FlatNonDefaultTags()（覆盖面零变）。baseline 删 2 行。register=no（gocell-internal-layout，
扫 adapters/mqtt，外部 vacuous）。行为零变更（verbatim MOVE + 恒等派生），无 //nolint。

ref: golang.org/x/tools/go/analysis Analyzer
